Create 2020-07-04-builder-pattern.md
Added below objects:
1. Human: Has the basic properties of the humans
2. Human Builder: Builder class to create the Human objects in very clean and neat way without adding extra frameworks to do the same
3. Relation: Relation is a binary relation object having left and right entity.
This is the initial simple builder design pattern as per the initial requirement. It takes the journey from the script pattern and other hazel involve in creating the objects with so many attributes to more clean way. Relation is added to take our example a bit ahead. Introducing some basic characters in our story. Our relation class and its descendant i.e the HusbandWife is the example of the inheritance. Here we have also established the relation between our characters. This could be directly related to what information we expect from our initial stage of the program. In short this example relates the design pattern thinking and the user requirements from the program.